move tip smt ast to separate file
